Move body parser to shared function
