+++ app/content_migration/management/commands/import_magazine_articles.py
@@ -1,16 +1,10 @@
-import re
-from typing import List
-
 import numpy as np
 import pandas as pd
-from bs4 import BeautifulSoup
-from bs4 import Tag as BS4_Tag
-from django.core.exceptions import ObjectDoesNotExist
-from django.core.management.base import BaseCommand, CommandError
-from tqdm import tqdm
-from wagtail.rich_text import RichText
 
-from contact.models import Meeting, Organization, Person
+from django.core.exceptions import ObjectDoesNotExist
+from django.core.management.base import BaseCommand
+from tqdm import tqdm
+
 from content_migration.management.commands.shared import (
     get_contact_from_author_data,
     get_existing_magazine_author_by_id,
@@ -22,85 +16,7 @@ from magazine.models import (
     MagazineIssue,
 )
 
-from .shared import parse_media_blocks
-
-
-def extract_pullquotes(item: str) -> List[str]:
-    """Get a list of all pullquote strings found within the item"""
-
-    return re.findall(r"\[pullquote\](.+?)\[\/pullquote\]", item.string)
-
-
-def clean_pullquote_tags(item: BS4_Tag) -> BS4_Tag:
-    """
-    Replace "[pullquote][/pullquote]" tags in string with "<span class='pullquote'></span>"
-    https://stackoverflow.com/a/44593228/1191545
-    """
-
-    replacement_values = [
-        ("[pullquote]", ""),
-        ("[/pullquote]", ""),
-    ]
-
-    for replacement_value in replacement_values:
-        item.string = item.string.replace(*replacement_value)
-
-    return item
-
-
-def parse_article_body_blocks(body):
-    article_body_blocks = []
-
-    try:
-        soup = BeautifulSoup(body, "html.parser")
-    except:
-        soup = False
-
-        print("Could not parse article body:", body)
-
-    # Placeholder for gathering successive items
-    rich_text_value = ""
-
-    if soup:
-        for item in soup:
-
-            item_has_value = item.string is not None
-
-            if item_has_value:
-
-                item_contains_pullquote = "pullquote" in item.string
-
-                if item_contains_pullquote:
-                    # Add current rich text value as rich text block, if not empty
-                    if rich_text_value != "":
-                        rich_text_block = ("rich_text", RichText(rich_text_value))
-
-                        article_body_blocks.append(rich_text_block)
-
-                        # reset rich text value
-                        rich_text_value = ""
-
-                    pullquotes = extract_pullquotes(item)
-
-                    # Add Pullquote block(s) to body streamfield
-                    # so they appear above the related rich text field
-                    # i.e. near the paragraph containing the pullquote
-                    for pullquote in pullquotes:
-                        block_content = ("pullquote", pullquote)
-
-                        article_body_blocks.append(block_content)
-
-                    item = clean_pullquote_tags(item)
-
-                rich_text_value += str(item)
-
-        if rich_text_value != "":
-            # Add Paragraph Block with remaining rich text elements
-            rich_text_block = ("rich_text", RichText(rich_text_value))
-
-            article_body_blocks.append(rich_text_block)
-
-    return article_body_blocks
+from .shared import parse_media_blocks, parse_body_blocks,
 
 
 def parse_article_authors(article, article_authors, magazine_authors_data):
@@ -180,7 +96,7 @@ class Command(BaseCommand):
             body_migrated = None
 
             if row["Body"] is not np.nan:
-                article_body_blocks = parse_article_body_blocks(row["Body"])
+                article_body_blocks = parse_body_blocks(row["Body"])
                 body_migrated = row["Body"]
 
             # Download and parse article media
